+++ Shared/ReflectionnaireTotal.cs
@@ -3,6 +3,6 @@
     public class CategoryTotal
     {
         public Category Category { get; set; }
-        public int TotalScore { get; set; }
+        public float TotalScore { get; set; }
     }
 }
